requirements.txt
requests==2.31.0
pyyaml==6.0.1
pytest==7.4.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.8.5
pandas==2.1.0
loguru==0.7.0